added trait and tests for lazy transforms
